Refactor function pxt
The computational steps in function pxt are hard to follow.
The underlying calculation is simple; pxt is the quotient of
l(x+t) and l(x);

  pxt = l(x+t) / l(x)

Complications arise when either argument, x or x + t, has a
fractional part and thus lies between two arguments for which we
know the values of l.  In this case we must interpolate between
these known values of l.

The helper function .interpolate does the interpolation and so
the definition of pxt is now much clearer and shorter.

A side benefit is that .interpolate can handle vector arguments;
therefore, pxt is now vectorized.